requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.23.0
openai-whisper==20230314
pycparser==2.21
regex==2023.8.8
requests==2.31.0
scipy==1.11.1
sounddevice==0.4.6
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
